Move uri logging processor out of postgres processor
